Added logic tier to backend and refactored code into controller - service - repository format
