--- AML.Server/Business/Media/IBorrowMediaProcessor.cs
@@ -1,8 +1,0 @@
-﻿using AML.Server.DTOs;
-
-namespace AML.Server.Business.Media;
-
-public interface IBorrowMediaProcessor
-{
-    Task Process(BorrowMediaRequest request);
-}
