--- AML.Server/Business/Media/ISearchMediaGetter.cs
@@ -1,8 +1,0 @@
-﻿using AML.Server.Models;
-
-namespace AML.Server.Business.Media;
-
-public interface ISearchMediaGetter
-{
-    Task<List<Models.Media>> Get(SearchMediaRequest? request);
-}
