+++ AML.Server/Controllers/OrderController.cs
@@ -13,26 +13,5 @@ namespace AML.Server.Controllers
         {
             _orderRepository = orderRepository;
         }
-
-        [HttpPost]
-        [Route("reserve-media")]
-        public async Task ReserveMedia(int mediaId, int accountId)
-        {
-            // add logic
-        }
-
-        [HttpPost]
-        [Route("borrow-media")]
-        public async Task BorrowMedia(int mediaId, int accountId, DateTime returnDate)
-        {
-            // add logic
-        }
-
-        [HttpPost]
-        [Route("return-media")]
-        public async Task ReturnMedia(int mediaId, int accountId, DateTime dateReturned)
-        {
-            // add logic
-        }
     }
 }
