+++ AML.Server/Interfaces/IMediaService.cs
@@ -3,9 +3,9 @@ using AML.Server.Models;
 
 namespace AML.Server.Interfaces
 {
-    public interface IMediaRepository
+    public interface IMediaService
     {
-        Task BorrowMedia(BorrowMediaRequest request);
         Task<List<Media>> GetMedia(SearchMediaRequest? request);
+        Task<bool> BorrowMedia(BorrowMediaRequest request);
     }
 }
