+++ AML.Server/Interfaces/IStockRepository.cs
@@ -2,7 +2,6 @@
 {
     public interface IStockRepository
     {
-        Task GetStockForBranch(int branchId, int mediaId, int amount);
-        Task RemoveMedia(int branchId, int mediaId);
+
     }
 }
