+++ AML.Server/Repositories/MediaRepository.cs
@@ -1,21 +1,49 @@
-﻿using AML.Server.Interfaces;
+﻿using AML.Server.Data;
+using AML.Server.DTOs;
+using AML.Server.Interfaces;
 using AML.Server.Models;
+using Microsoft.EntityFrameworkCore;
 
 namespace AML.Server.Repositories
 {
     public class MediaRepository : IMediaRepository
     {
-        // Add a dbcontext here
+        private readonly DBContext dbContext;
 
-        public MediaRepository(/*inject dbcontext*/)
+        public MediaRepository(DBContext dbContext)
         {
-            /*Assign injected context to declared above*/
+            this.dbContext = dbContext;
         }
 
-        public async Task<List<Media>> SearchMedia(string searchTerm)
+        public async Task<List<Models.Media>> GetMedia(SearchMediaRequest? request)
         {
-            // add logic
-            return null;
+            var mediaList = new List<Models.Media>();
+
+            if (request.Search is null)
+            {
+                mediaList = await this.dbContext.Media.ToListAsync();
+            }
+            else
+            {
+                mediaList = await this.dbContext.Media.Where(x => x.Name.ToLower().Trim().Contains(request.Search.ToLower().Trim())).ToListAsync();
+            }
+
+            return mediaList;
+        }
+
+        public async Task BorrowMedia(BorrowMediaRequest request)
+        {
+            var order = new Order()
+            {
+                AccountId = request.AccountId,
+                MediaId = request.MediaId,
+                OrderDate = request.Start,
+                ReturnDate = request.End,
+                OrderType = OrderType.Borrow
+            };
+
+            this.dbContext.Orders.Add(order);
+            await this.dbContext.SaveChangesAsync();
         }
     }
 }
